+++ raiden/utils/test_support/synapse_config.yaml.template
