+++ requirements/requirements-ci.txt
@@ -5,7 +5,7 @@
 #    './deps compile' (for details see README)
 #
 alabaster==0.7.12
-altgraph==0.16.1          # via macholib, pyinstaller
+altgraph==0.16.1          # via pyinstaller
 aniso8601==7.0.0
 apipkg==1.5
 appdirs==1.4.3
@@ -14,11 +14,11 @@ astroid==2.2.5
 astunparse==1.6.2
 attrdict==2.0.1
 attrs==19.3.0
-automat==0.7.0
+automat==20.2.0
 autopep8==1.4.4
 babel==2.7.0
 base58==2.0.0
-bcrypt==3.1.6
+bcrypt==3.1.7
 black==19.3b0
 bleach==3.1.1
 bump2version==0.5.10
@@ -78,7 +78,6 @@ jinja2==2.10.1
 jsonschema==3.2.0
 lazy-object-proxy==1.4.1
 lru-dict==1.1.6
-macholib==1.14            # via pyinstaller
 markupsafe==1.1.1
 marshmallow-dataclass==6.0.0
 marshmallow-enum==1.5.1
@@ -89,7 +88,7 @@ matrix-synapse==1.5.1
 mccabe==0.6.1
 mirakuru==2.1.2
 more-itertools==7.0.0
-msgpack==0.6.1
+msgpack==1.0.0
 multiaddr==0.0.9
 mypy-extensions==0.4.3
 mypy==0.761
@@ -101,12 +100,12 @@ packaging==19.0
 parsimonious==0.8.1
 pdbpp==0.10.0
 pexpect==4.7.0
-phonenumbers==8.10.13
+phonenumbers==8.11.5
 pickleshare==0.7.5
-pillow==6.2.0
+pillow==7.0.0
 pip-tools==4.4.0
 pluggy==0.12.0
-prometheus-client==0.3.1
+prometheus-client==0.7.1
 protobuf==3.11.3
 psutil==5.6.3
 ptyprocess==0.6.0
@@ -115,19 +114,19 @@ py-geth==2.1.0
 py-solc==3.2.0
 py-spy==0.3.2
 py==1.8.0
-pyasn1-modules==0.2.5
-pyasn1==0.4.5
+pyasn1-modules==0.2.8
+pyasn1==0.4.8
 pycodestyle==2.5.0
 pycparser==2.19
 pycryptodome==3.8.2
 pyflakes==2.1.1
 pygments==2.4.2
-pyhamcrest==1.9.0
+pyhamcrest==2.0.2
 pyinstaller==3.6
 pylint==2.3.1
 pymacaroons==0.13.0
 pynacl==1.3.0
-pyopenssl==19.0.0
+pyopenssl==19.1.0
 pyparsing==2.4.0
 pyrsistent==0.15.7
 pysha3==1.0.2
@@ -140,7 +139,7 @@ python-dateutil==2.8.0    # via s3cmd
 python-magic==0.4.15      # via s3cmd
 pytoml==0.1.20
 pytz==2019.1
-pyyaml==5.1.1
+pyyaml==5.3
 raiden-contracts==0.36.1
 raiden-webui==0.11.0
 readme-renderer==24.0
@@ -152,9 +151,9 @@ s3cmd==2.0.2
 semantic-version==2.6.0
 semver==2.8.1
 service-identity==18.1.0
-signedjson==1.0.0
+signedjson==1.1
 simplegeneric==0.8.1
-simplejson==3.16.0
+simplejson==3.17.0
 six==1.12.0
 snowballstemmer==1.2.1
 sortedcontainers==2.1.0
@@ -169,7 +168,7 @@ toml==0.10.0
 toolz==0.9.0
 traitlets==4.3.2
 treq==18.6.0
-twisted[tls]==19.2.1
+twisted[tls]==19.10.0
 typed-ast==1.4.0
 typing-extensions==3.7.4.1
 typing-inspect==0.4.0
@@ -186,7 +185,7 @@ wheel==0.33.4
 wmctrl==0.3
 wrapt==1.11.1
 zipp==2.2.0
-zope.interface==4.6.0
+zope.interface==4.7.1
 
 # The following packages are considered to be unsafe in a requirements file:
 # setuptools
