+++ requirements/requirements-dev.txt
@@ -13,16 +13,16 @@ astroid==2.2.5            # via pylint
 astunparse==1.6.2
 attrdict==2.0.1
 attrs==19.3.0
-automat==0.7.0            # via twisted
+automat==20.2.0
 autopep8==1.4.4
 babel==2.7.0
 base58==2.0.0
-bcrypt==3.1.6             # via matrix-synapse
+bcrypt==3.1.7
 black==19.3b0
-bleach==3.1.1             # via matrix-synapse, readme-renderer
+bleach==3.1.1
 bump2version==0.5.10
 cachetools==3.1.1
-canonicaljson==1.1.4      # via matrix-synapse, signedjson
+canonicaljson==1.1.4
 certifi==2019.3.9
 cffi==1.12.3
 chardet==3.0.4
@@ -30,11 +30,11 @@ click==7.0
 coincurve==13.0.0
 colorama==0.4.1
 colour==0.1.5
-constantly==15.1.0        # via twisted
+constantly==15.1.0
 coverage==4.5.3
-cryptography==2.8         # via pyopenssl, service-identity
+cryptography==2.8
 cytoolz==0.10.1
-daemonize==2.5.0          # via matrix-synapse
+daemonize==2.5.0
 decorator==4.4.0
 docutils==0.14
 entrypoints==0.3          # via flake8
@@ -55,19 +55,19 @@ flake8==3.7.9
 flask-cors==3.0.8
 flask-restful==0.3.7
 flask==1.0.3
-frozendict==1.2           # via canonicaljson, matrix-synapse
+frozendict==1.2
 gevent==1.5a3
 graphviz==0.13.2
 greenlet==0.4.15
 grequests==0.3.0
 guppy3==3.0.9
 hexbytes==0.2.0
-hyperlink==19.0.0         # via twisted
+hyperlink==19.0.0
 hypothesis==4.24.3
 idna==2.8
 imagesize==1.1.0
 importlib-metadata==1.5.0
-incremental==17.5.0       # via treq, twisted
+incremental==17.5.0
 ipfshttpclient==0.4.12
 ipython-genutils==0.2.0
 ipython==4.2.1
@@ -87,11 +87,11 @@ matrix-synapse==1.5.1
 mccabe==0.6.1             # via flake8, pylint
 mirakuru==2.1.2
 more-itertools==7.0.0     # via pytest
-msgpack==0.6.1            # via matrix-synapse
+msgpack==1.0.0
 multiaddr==0.0.9
-netaddr==0.7.19
 mypy-extensions==0.4.3
 mypy==0.761
+netaddr==0.7.19
 netifaces==0.10.9
 networkx==2.3
 objgraph==3.4.1
@@ -99,12 +99,12 @@ packaging==19.0
 parsimonious==0.8.1
 pdbpp==0.10.0
 pexpect==4.7.0
-phonenumbers==8.10.13     # via matrix-synapse
+phonenumbers==8.11.5
 pickleshare==0.7.5
-pillow==6.2.0             # via matrix-synapse
+pillow==7.0.0
 pip-tools==4.4.0
 pluggy==0.12.0            # via pytest
-prometheus-client==0.3.1  # via matrix-synapse
+prometheus-client==0.7.1
 protobuf==3.11.3
 psutil==5.6.3
 ptyprocess==0.6.0
@@ -113,18 +113,18 @@ py-geth==2.1.0
 py-solc==3.2.0
 py-spy==0.3.2
 py==1.8.0                 # via pytest
-pyasn1-modules==0.2.5     # via matrix-synapse, service-identity
-pyasn1==0.4.5             # via matrix-synapse, pyasn1-modules, service-identity
+pyasn1-modules==0.2.8
+pyasn1==0.4.8
 pycodestyle==2.5.0        # via autopep8, flake8
 pycparser==2.19
 pycryptodome==3.8.2
 pyflakes==2.1.1           # via flake8
 pygments==2.4.2
-pyhamcrest==1.9.0         # via twisted
+pyhamcrest==2.0.2
 pylint==2.3.1
-pymacaroons==0.13.0       # via matrix-synapse
-pynacl==1.3.0             # via matrix-synapse, pymacaroons, signedjson
-pyopenssl==19.0.0         # via matrix-synapse, twisted
+pymacaroons==0.13.0
+pynacl==1.3.0
+pyopenssl==19.1.0
 pyparsing==2.4.0
 pyrsistent==0.15.7
 pysha3==1.0.2
@@ -135,7 +135,7 @@ pytest-xdist==1.31.0
 pytest==5.3.5
 pytoml==0.1.20
 pytz==2019.1
-pyyaml==5.1.1             # via matrix-synapse
+pyyaml==5.3
 raiden-contracts==0.36.1
 raiden-webui==0.11.0
 readme-renderer==24.0
@@ -145,13 +145,13 @@ responses==0.10.9
 rlp==1.1.0
 semantic-version==2.6.0
 semver==2.8.1
-service-identity==18.1.0  # via matrix-synapse, twisted
-signedjson==1.0.0         # via matrix-synapse
+service-identity==18.1.0
+signedjson==1.1
 simplegeneric==0.8.1
-simplejson==3.16.0        # via canonicaljson
+simplejson==3.17.0
 six==1.12.0
 snowballstemmer==1.2.1
-sortedcontainers==2.1.0   # via matrix-synapse
+sortedcontainers==2.1.0
 sphinx-rtd-theme==0.4.3
 sphinx==1.7.9
 sphinxcontrib-httpdomain==1.7.0
@@ -162,25 +162,25 @@ structlog==19.1.0
 toml==0.10.0              # via black
 toolz==0.9.0
 traitlets==4.3.2
-treq==18.6.0              # via matrix-synapse
-twisted[tls]==19.2.1      # via matrix-synapse, treq
+treq==18.6.0
+twisted[tls]==19.10.0
 typed-ast==1.4.0          # via astroid, mypy
 typing-extensions==3.7.4.1
 typing-inspect==0.4.0
-unpaddedbase64==1.1.0     # via matrix-synapse, signedjson
+unpaddedbase64==1.1.0
 urllib3==1.25.3
 varint==1.0.2
 wcwidth==0.1.7            # via pytest
 web3==5.5.1
 webargs==5.3.1
-webencodings==0.5.1       # via bleach
+webencodings==0.5.1
 websockets==8.1
 werkzeug==0.15.4
 wheel==0.33.4
 wmctrl==0.3               # via pdbpp
 wrapt==1.11.1             # via astroid
 zipp==2.2.0
-zope.interface==4.6.0     # via twisted
+zope.interface==4.7.1
 
 # The following packages are considered to be unsafe in a requirements file:
 # setuptools
