+++ requirements/requirements.txt
@@ -7,16 +7,23 @@
 aniso8601==7.0.0          # via flask-restful
 asn1crypto==0.24.0        # via coincurve
 attrdict==2.0.1           # via eth-account
-attrs==19.3.0             # via jsonschema
+attrs==19.3.0             # via automat, jsonschema, matrix-synapse, service-identity, treq, twisted
+automat==20.2.0           # via twisted
 base58==2.0.0             # via multiaddr
+bcrypt==3.1.7             # via matrix-synapse
+bleach==3.1.1             # via matrix-synapse
 cachetools==3.1.1
+canonicaljson==1.1.4      # via matrix-synapse, signedjson
 certifi==2019.3.9         # via requests
-cffi==1.12.3              # via coincurve
+cffi==1.12.3              # via bcrypt, coincurve, cryptography, pynacl
 chardet==3.0.4            # via requests
 click==7.0                # via flask, raiden-contracts
 coincurve==13.0.0
 colorama==0.4.1
+constantly==15.1.0        # via twisted
+cryptography==2.8         # via pyopenssl, service-identity
 cytoolz==0.10.1           # via eth-keyfile, eth-utils
+daemonize==2.5.0          # via matrix-synapse
 decorator==4.4.0          # via ipython, networkx, traitlets
 eth-abi==2.1.0            # via eth-account, raiden-contracts, web3
 eth-account==0.4.0        # via web3
@@ -30,19 +37,22 @@ filelock==3.0.12
 flask-cors==3.0.8
 flask-restful==0.3.7
 flask==1.0.3
+frozendict==1.2           # via canonicaljson, matrix-synapse
 gevent==1.5a3
 graphviz==0.13.2          # via objgraph
 greenlet==0.4.15          # via gevent
 guppy3==3.0.9
 hexbytes==0.2.0           # via eth-account, eth-rlp, web3
-idna==2.8                 # via requests
+hyperlink==19.0.0         # via twisted
+idna==2.8                 # via hyperlink, matrix-synapse, requests, twisted
 importlib-metadata==1.5.0  # via jsonschema
+incremental==17.5.0       # via treq, twisted
 ipfshttpclient==0.4.12    # via web3
 ipython-genutils==0.2.0   # via traitlets
 ipython==4.2.1
 itsdangerous==1.1.0       # via flask
-jinja2==2.10.1            # via flask
-jsonschema==3.2.0         # via web3
+jinja2==2.10.1            # via flask, matrix-synapse
+jsonschema==3.2.0         # via matrix-synapse, web3
 lru-dict==1.1.6           # via web3
 markupsafe==1.1.1         # via jinja2
 marshmallow-dataclass==6.0.0
@@ -50,48 +60,69 @@ marshmallow-enum==1.5.1
 marshmallow-polyfield==5.8
 marshmallow==3.4.0
 matrix-client==0.3.2
+matrix-synapse==1.5.1
 mirakuru==2.1.2
+msgpack==1.0.0            # via matrix-synapse
 multiaddr==0.0.9          # via ipfshttpclient
-netaddr==0.7.19           # via multiaddr
-mypy-extensions==0.4.3    # via raiden-contracts, typing-inspect
+mypy-extensions==0.4.3    # via typing-inspect
+netaddr==0.7.19           # via matrix-synapse, multiaddr
 netifaces==0.10.9
 networkx==2.3
 objgraph==3.4.1
 parsimonious==0.8.1       # via eth-abi
 pexpect==4.7.0            # via ipython
+phonenumbers==8.11.5      # via matrix-synapse
 pickleshare==0.7.5        # via ipython
+pillow==7.0.0             # via matrix-synapse
+prometheus-client==0.7.1  # via matrix-synapse
 protobuf==3.11.3          # via web3
 psutil==5.6.3
 ptyprocess==0.6.0         # via pexpect
 py-ecc==1.4.7             # via raiden-contracts
 py-geth==2.1.0
 py-solc==3.2.0
+pyasn1-modules==0.2.8     # via matrix-synapse, service-identity
+pyasn1==0.4.8             # via matrix-synapse, pyasn1-modules, service-identity
 pycparser==2.19           # via cffi
 pycryptodome==3.8.2       # via eth-hash, eth-keyfile
+pyhamcrest==2.0.2         # via twisted
+pymacaroons==0.13.0       # via matrix-synapse
+pynacl==1.3.0             # via matrix-synapse, pymacaroons, signedjson
+pyopenssl==19.1.0         # via matrix-synapse, twisted
 pyrsistent==0.15.7        # via jsonschema
 pysha3==1.0.2
 pytoml==0.1.20
 pytz==2019.1              # via flask-restful
+pyyaml==5.3               # via matrix-synapse
 raiden-contracts==0.36.1
 raiden-webui==0.11.0
 requests==2.22.0
 rlp==1.1.0                # via eth-account, eth-rlp, raiden-contracts
 semantic-version==2.6.0   # via py-geth, py-solc
 semver==2.8.1             # via raiden-contracts
+service-identity==18.1.0  # via matrix-synapse, twisted
+signedjson==1.1           # via matrix-synapse
 simplegeneric==0.8.1      # via ipython
-six==1.12.0               # via attrdict, flask-cors, flask-restful, ipfshttpclient, jsonschema, multiaddr, parsimonious, protobuf, pyrsistent, structlog, traitlets
+simplejson==3.17.0        # via canonicaljson
+six==1.12.0               # via attrdict, automat, bcrypt, bleach, canonicaljson, cryptography, flask-cors, flask-restful, ipfshttpclient, jsonschema, matrix-synapse, multiaddr, parsimonious, protobuf, pymacaroons, pynacl, pyopenssl, pyrsistent, structlog, traitlets, treq
+sortedcontainers==2.1.0   # via matrix-synapse
 structlog==19.1.0
 toolz==0.9.0              # via cytoolz
 traitlets==4.3.2          # via ipython
+treq==18.6.0              # via matrix-synapse
+twisted[tls]==19.10.0     # via matrix-synapse, treq
 typing-extensions==3.7.4.1
 typing-inspect==0.4.0     # via marshmallow-dataclass
+unpaddedbase64==1.1.0     # via matrix-synapse, signedjson
 urllib3==1.25.3           # via requests
 varint==1.0.2             # via multiaddr
 web3==5.5.1
 webargs==5.3.1
+webencodings==0.5.1       # via bleach
 websockets==8.1           # via web3
 werkzeug==0.15.4          # via flask
 zipp==2.2.0               # via importlib-metadata
+zope.interface==4.7.1     # via twisted
 
 # The following packages are considered to be unsafe in a requirements file:
 # setuptools
